requirements.txt
pygame==2.0.0
numpy==1.18.5
tensorflow==2.3.1
pillow==8.0.1
sphinx==3.3.1
sphinx_rtd_theme==0.5.0